Bug fixes and improvements in Numba interfaces
I have fixed a number of bugs on the installation, as well as
fixed the Numba interfaces. The JIT should cache the compilation
and make running the model quite fast. I have left PROSPECT in
pure python (not numba annotations) as there doesn't seem to be
much point in doing that.

The thermal interface isn't tested yet.